# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Dungeon Master API
  version: 2.0.0
  description: |
    HTTP API for the AI Dungeon Master engine.

    The **v2** API (`/v2/*`) wraps every response in a typed, versioned
    `Envelope` — `{ type, version, payload, requestId }` — so native clients get
    a stable, self-describing contract. The legacy `/api/game/*` endpoints remain
    for existing clients and are documented under the `legacy` tag.
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: v2
    description: Versioned, envelope-wrapped API.
  - name: legacy
    description: Original unversioned API (kept for backward compatibility).
paths:
  /v2/status:
    get:
      tags: [v2]
      operationId: getStatusV2
      summary: Current game status as a typed envelope.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Game status envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusEnvelope'
  /v2/action:
    post:
      tags: [v2]
      operationId: submitActionV2
      summary: Apply a choice; returns the updated game status envelope.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionRequest'
      responses:
        '200':
          description: Updated game status envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusEnvelope'
        '400':
          description: Missing or unknown choice.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v2/narrate:
    post:
      tags: [v2]
      operationId: narrateV2
      summary: Generate a dungeon-master narration via the active LLM provider.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NarrateRequest'
      responses:
        '200':
          description: Narration envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NarrativeEnvelope'
  /v2/save:
    post:
      tags: [v2]
      operationId: saveGameV2
      summary: Persist the caller's game engine to a session-scoped save file.
      description: |
        Authenticated callers save under `game.saves.dir/{sessionId}.json`.
        Unauthenticated callers share the process-default engine and save as
        `default.json`. Each authenticated session has its own engine instance.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Save acknowledgement envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSaveEnvelope'
  /v2/load:
    post:
      tags: [v2]
      operationId: loadGameV2
      summary: Restore the caller's game engine from its save file.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Restored game status envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusEnvelope'
        '400':
          description: No save file for this caller.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v2/reset:
    post:
      tags: [v2]
      operationId: resetGameV2
      summary: Start a fresh engine for the caller (new party/quest).
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Fresh game status envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusEnvelope'
  /v2/catalog:
    get:
      tags: [v2]
      operationId: getCatalogV2
      summary: Installed content packs and registered plugins (mod browser).
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Catalog envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogEnvelope'
  /v2/catalog/packs:
    post:
      tags: [v2]
      operationId: uploadPackV2
      summary: Upload and install a content-pack zip at runtime; returns the updated catalog.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: replace
          in: query
          required: false
          schema: { type: boolean, default: false }
          description: Overwrite an already-installed pack with the same id.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: >-
                    Pack zip — pack.yaml plus optional items/, monsters/,
                    strings/, quests/, campaigns/, npcs/, factions/. Pure data;
                    code-bearing mods use the plugin loader instead.
      responses:
        '201':
          description: Pack installed; updated catalog envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogEnvelope'
        '200':
          description: Existing pack replaced; updated catalog envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogEnvelope'
        '400':
          description: Invalid upload (bad zip, unsafe paths, missing/invalid pack.yaml, size limits).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Pack id already installed and replace=false.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v2/catalog/packs/{id}/enable:
    post:
      tags: [v2]
      operationId: enablePackV2
      summary: Enable a content pack; returns the updated catalog.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated catalog envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogEnvelope'
        '404':
          description: Unknown content pack.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v2/catalog/packs/{id}/disable:
    post:
      tags: [v2]
      operationId: disablePackV2
      summary: Disable a content pack; returns the updated catalog.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated catalog envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogEnvelope'
        '404':
          description: Unknown content pack.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v2/entitlements/verify:
    post:
      tags: [v2]
      operationId: verifyReceiptV2
      summary: Validate a purchase receipt via its storefront and grant the entitlement.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyReceiptRequest'
      responses:
        '200':
          description: Receipt valid; entitlement granted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementEnvelope'
        '402':
          description: Receipt failed verification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementEnvelope'
        '401':
          description: No authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v2/entitlements:
    get:
      tags: [v2]
      operationId: listEntitlementsV2
      summary: List the caller's owned products.
      parameters:
        - $ref: '#/components/parameters/XRequestId'
      responses:
        '200':
          description: Entitlements envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementEnvelope'
        '401':
          description: No authenticated session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/game/status:
    get:
      tags: [legacy]
      operationId: getStatusLegacy
      summary: Legacy game status (flat pipe-delimited party summary string).
      responses:
        '200':
          description: Game status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusResponse'
  /api/game/choices:
    get:
      tags: [legacy]
      operationId: getChoicesLegacy
      summary: Available choice labels.
      responses:
        '200':
          description: Choice labels.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/game/action:
    post:
      tags: [legacy]
      operationId: submitActionLegacy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionRequest'
      responses:
        '200':
          description: Acceptance message.
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Missing or unknown choice.
          content:
            text/plain:
              schema:
                type: string
  /api/game/start:
    post:
      tags: [legacy]
      operationId: startQuestLegacy
      summary: (Re)start the quest and return the opening banner.
      responses:
        '200':
          description: Opening banner text.
          content:
            text/plain:
              schema:
                type: string
  /api/game/save:
    post:
      tags: [legacy]
      operationId: saveGameLegacy
      responses:
        '200':
          description: Confirmation message.
          content:
            text/plain:
              schema:
                type: string
  /api/game/load:
    post:
      tags: [legacy]
      operationId: loadGameLegacy
      responses:
        '200':
          description: Confirmation message.
          content:
            text/plain:
              schema:
                type: string
components:
  parameters:
    XRequestId:
      name: X-Request-Id
      in: header
      required: false
      description: >
        Optional correlation id echoed back in the response envelope's
        requestId. A server-generated UUID is used when omitted.
      schema:
        type: string
  schemas:
    Envelope:
      type: object
      description: Base transport envelope. Concrete responses specialize `payload`.
      required: [type, version, payload, requestId]
      properties:
        type:
          type: string
          description: Payload discriminator (game_status, narrative_update, error).
        version:
          type: integer
          description: Schema version of the payload contract.
          example: 1
        payload:
          type: object
          description: Typed body; shape depends on `type`.
        requestId:
          type: string
          description: Correlation id (echoed from X-Request-Id or generated).
    GameStatusEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            type:
              type: string
              example: game_status
            payload:
              $ref: '#/components/schemas/GameStatusV2'
    NarrativeEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            type:
              type: string
              example: narrative_update
            payload:
              $ref: '#/components/schemas/NarrativePayload'
    GameSaveEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            type:
              type: string
              example: game_save
            payload:
              type: object
              properties:
                saved: { type: boolean }
                path: { type: string }
                sessionScoped: { type: boolean }
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            type:
              type: string
              example: error
            payload:
              $ref: '#/components/schemas/ErrorPayload'
    CatalogEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            type:
              type: string
              example: catalog
            payload:
              $ref: '#/components/schemas/CatalogPayload'
    EntitlementEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties:
            type:
              type: string
              example: entitlement
            payload:
              $ref: '#/components/schemas/EntitlementPayload'
    MemberState:
      type: object
      description: Public snapshot of a single party member.
      properties:
        name: { type: string }
        role: { type: string }
        level: { type: integer }
        hp: { type: integer }
        maxHp: { type: integer }
        mana: { type: integer }
        maxMana: { type: integer }
        armorClass: { type: integer }
        alive: { type: boolean }
        ascendant: { type: boolean }
        statuses:
          type: array
          items: { type: string }
    GameStatusV2:
      type: object
      properties:
        party:
          type: array
          items:
            $ref: '#/components/schemas/MemberState'
        chaosLevel: { type: integer }
        combatActive: { type: boolean }
        availableChoices:
          type: array
          items: { type: string }
        recentHistory:
          type: array
          items: { type: string }
        quest:
          $ref: '#/components/schemas/QuestInfo'
        recentEvents:
          description: >-
            Compact story-memory facts from the engine's Chronicle (newest
            last), e.g. "Quest completed: The Weeping Tree". Bounded server-side.
          type: array
          items: { type: string }
        location:
          description: Current party location from the engine WorldMap.
          type: string
        discoveredRifts:
          description: Named rifts the party has discovered (starting set + completed quests).
          type: array
          items: { type: string }
    QuestInfo:
      description: Outcome snapshot of the current quest.
      type: object
      nullable: true
      properties:
        title: { type: string }
        completed: { type: boolean }
        failed: { type: boolean }
        progress:
          description: Scene progress in [0,1].
          type: number
          format: double
    NarrativePayload:
      type: object
      properties:
        text: { type: string }
        provider:
          type: string
          description: Active provider id, e.g. local-stub.
        tokensUsed: { type: integer }
        fallback:
          type: boolean
          description: True when a guardrail/fallback response was returned.
    CatalogPayload:
      type: object
      description: Snapshot of everything installed and dispatchable in the engine.
      properties:
        contentPacks:
          type: array
          items:
            $ref: '#/components/schemas/PackInfo'
        plugins:
          $ref: '#/components/schemas/PluginSummary'
        narration:
          $ref: '#/components/schemas/NarrationInfo'
    PackInfo:
      type: object
      description: One installed content pack.
      properties:
        id: { type: string }
        displayName: { type: string }
        version: { type: string }
        monsters: { type: integer }
        items: { type: integer }
        enabled: { type: boolean }
    PluginSummary:
      type: object
      description: Registered plugin ids (or biome tags) grouped by SPI.
      properties:
        spellEffects: { type: array, items: { type: string } }
        itemEffects: { type: array, items: { type: string } }
        encounterBiomes: { type: array, items: { type: string } }
        lootBiomes: { type: array, items: { type: string } }
        questScripts: { type: array, items: { type: string } }
        storefronts: { type: array, items: { type: string } }
        llmProviders: { type: array, items: { type: string } }
    NarrationInfo:
      type: object
      description: Active narration provider and everything available to switch to.
      properties:
        active: { type: string }
        health:
          type: string
          description: Provider health (OK, DEGRADED, DOWN, UNKNOWN).
        available:
          type: array
          items: { type: string }
    EntitlementPayload:
      type: object
      properties:
        granted: { type: boolean }
        productId: { type: string, nullable: true }
        storefront: { type: string, nullable: true }
        reason: { type: string }
        owned:
          type: array
          items: { type: string }
    VerifyReceiptRequest:
      type: object
      required: [productId, receipt]
      properties:
        storefront: { type: string, example: dev }
        productId: { type: string, example: sku_gold }
        receipt: { type: string }
    ErrorPayload:
      type: object
      properties:
        message: { type: string }
    ActionRequest:
      type: object
      required: [choiceLabel]
      properties:
        choiceLabel:
          type: string
          example: Attack
    NarrateRequest:
      type: object
      properties:
        prompt:
          type: string
          example: I search the altar for traps
    GameStatusResponse:
      type: object
      description: Legacy flat status; `partySummary` is a pipe-delimited string.
      properties:
        partySummary: { type: string }
        chaosLevel: { type: integer }
        combatActive: { type: boolean }
        availableChoices:
          type: array
          items: { type: string }
        recentHistory:
          type: array
          items: { type: string }